ejercicio 1 parte 2 completa

+++ Obligatorio1_MDA/src/logica/Producto.java
@@ -41,7 +41,7 @@ public class Producto {
         this.proveedor = proveedor;
     }
     
-    public float getPrecio() {
+    public int getPrecio() {
         return precio;
     }
